ai_dev_team/pyproject.toml
[project]
name = "ai-dev-team"
version = "1.0.0"
description = "AI-powered development team using CrewAI with local models"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.80.0",
    "ollama>=0.3.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
dev-team = "ai_dev_team.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
